Split buildings.rs (593 lines) into buildings/ subdirectory
Split crates/simulation/src/buildings.rs into focused modules:
- types.rs: Building, MixedUseBuilding, UnderConstruction structs and capacity logic
- spawning.rs: BuildingSpawnTimer, EligibleCells, spawn/eligibility systems
- construction.rs: progress_construction system
- plugin.rs: BuildingsPlugin registration
- tests.rs: all unit tests
- mod.rs: thin facade with re-exports

All public items re-exported from mod.rs — no external import changes needed.
